Refactor: Move user action handling to view
The view is now responsible for catching user interactions with the UI
(i.e. clicks on cells). The controller - which is responsible for
orchestrating things between the view and the model - registers a
listener to the view. The listener is executed by the view to notify
the controller the player is shooting at a cell. The controllers then
tells the domain (via our model) to handle that event.

The controller (application layer) is now simply listening to some
events on the user interface (presentation) and delegating it to the
model (domain).

What we learned:
* Presentation, application and domain logic separation
* Registering a simple callback