Refactor: Decouple acceptance tests (work in progress)
Each test *should* be independant. Even acceptance tests. We expect
the system under test to be in a clean blank state when a test is
starting to reduce risks of flickering and to ensure the deterministic
nature of our tests.

In order to reach that situation, we start a new game before every
single end-to-end tests.

What we learned:
* All tests should be independent
* Tests start with a system in a clean (or very well-known) state